Fix Netrw window focus when g:netrw_liststyle = 3
If the Netrw plugin is configured in tree style listing mode
(i.e. g:netrw_liststyle = 3), then a buffer named 'NetrwTreeListing' is shown
(i.e. active window).

This happens only after the original buffer, which has the directory as its
name, is focused for the first time.

We work around this by comparing the 'netrw_curdir' variable of the
'NetrwTreeListing' buffer against the directory name.